Add routers
Replace routing written in main.go with a separate package dedicated to
managing http.ServeMuxes. This way, we keep main.go nice and clean,
clear of uglier implementation details.

To make this addition clean, this commit also adds a `static_prefix'
global configuration flag, so we're not hardcoding the static path
anywhere.