Remove bower, and download highlight.js instead
Download highlight.js from highlightjs.org with curl and a shell script
instead of using the bower package, in which highlight.pack.js was
nearly a megabyte(!!!) since it included all supported languages.
Instead, this commit swaps to a default, short list of langauges that
users can modify as needed.

This reverts commit 3820f4a66ca84f2dd545224310ce0f5602344327.